backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6

# Twilio
twilio>=8.10.0

# AI/ML
groq>=0.5.0
sentence-transformers>=2.3.1
faiss-cpu>=1.9.0
langdetect>=1.0.9

# Audio processing
pydub>=0.25.1

# Utilities
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
